cleaned up organic job flow test
